app/etl/requirements.txt
elasticsearch==8.5.1
pandas==1.5.3
openpyxl==3.1.2
numpy==1.23.5
python-dateutil==2.8.2
python-dotenv==1.0.0
loguru==0.6.0